Refactor mapping logic
