+++ app/src/main/java/nz/eloque/foss_wallet/model/PassWithLocalization.kt
@@ -4,7 +4,7 @@ import androidx.room.Embedded
 import androidx.room.Relation
 import nz.eloque.foss_wallet.model.field.PassContent
 import nz.eloque.foss_wallet.model.field.PassField
-import java.util.Locale
+import nz.eloque.foss_wallet.utils.toMapping
 
 private const val CHANGE_MESSAGE_FORMAT = "%@"
 
@@ -18,7 +18,7 @@ data class PassWithLocalization(
     val localizations: List<PassLocalization>,
 ) {
     fun applyLocalization(locale: String): Pass {
-        val mapping = localeMapping(locale)
+        val mapping = localizations.toMapping(locale)
         return pass.copy(
             description = mapping[pass.description]?.text ?: pass.description,
             headerFields = pass.headerFields.applyLocalization(mapping),
@@ -49,22 +49,4 @@ data class PassWithLocalization(
         } else {
             this
         }
-
-    private fun localeMapping(locale: String): Map<String, PassLocalization> {
-        val availableLanguageTags = localizations.map { it.lang.normalizedLanguageTag() }.distinct()
-        val preferredLanguages = listOf(locale, "en")
-        val bestMatch =
-            Locale.lookupTag(
-                preferredLanguages.mapNotNull { it.normalizedLanguageTag().toLanguageRangeOrNull() },
-                availableLanguageTags,
-            ) ?: "en"
-
-        return localizations
-            .filter { it.lang.normalizedLanguageTag().equals(bestMatch, ignoreCase = true) }
-            .associateBy { it.label }
-    }
-
-    private fun String.normalizedLanguageTag(): String = replace('_', '-')
-
-    private fun String.toLanguageRangeOrNull(): Locale.LanguageRange? = runCatching { Locale.LanguageRange(this) }.getOrNull()
 }
